==> 02_block_node_pairing.ipynb <==
import pickle
import numpy as np
import matplotlib.pyplot as plt
import networkx as nx
import geopandas as gpd
import pandas as pd
import numpy as np
import pickle
from shapely import Point
from shapely.ops import nearest_points
from rtree import index
from tqdm import tqdm

from shapely.geometry import Point, Polygon, LineString,MultiLineString

# import Node_Block.pkl

with open(r'intermediate_files/Node_Block.pkl', 'rb') as f:
    Node_Block = pickle.load(f)

Node_Block

# number of unique values in Node_Block['Node']

len(Node_Block['Node'].unique())

with open(r'nodes_edges_ucla_big_graph.pickle', 'rb') as handle:
    B_matrix_sliced,B_matrix_str_sliced,nodes_coordinates_array = pickle.load(handle)

# dimensions of nodes_coordinates_array

nodes_coordinates_array.shape

Origin_Destination = pd.read_csv('lodes_od_data/ca_od_main_JT00_2021.csv')

Initial_Origin_Destination_Count = len(Origin_Destination)

# Create dictionary for block and node pairings (identified from BlockDetermination.ipynb)
block_node_dict = {}

for _, row in Node_Block.iterrows():
    node_id = row['Node']
    block_id = row['BlockNumber']
    # block_coord = row['BlockBoundary']
    if int(block_id) not in block_node_dict:
        block_node_dict[int(block_id)] = []
    block_node_dict[int(block_id)].append(node_id)

Origin_Destination_Node_Added = Origin_Destination.copy()

Origin_Destination_Node_Added

# drop createdate column

Origin_Destination_Node_Added = Origin_Destination_Node_Added.drop(columns=['createdate'])

# Map nodes to work and home blocks
Origin_Destination_Node_Added['w_node_id'] = Origin_Destination_Node_Added['w_geocode'].map(block_node_dict)
Origin_Destination_Node_Added['h_node_id'] = Origin_Destination_Node_Added['h_geocode'].map(block_node_dict)

Origin_Destination_Node_Added

# how many rows have w_node_id and h_node_id as NaN

Origin_Destination_Node_Added['w_node_id'].isna().sum() / len(Origin_Destination_Node_Added)

# Remove rows where the node columns are missing (i.e. the block is not in the node dictionary)
Origin_Destination_Node_Added = Origin_Destination_Node_Added[Origin_Destination_Node_Added['w_node_id'].notnull()]
Origin_Destination_Node_Added = Origin_Destination_Node_Added[Origin_Destination_Node_Added['h_node_id'].notnull()]

Origin_Destination_Node_Added.info(memory_usage='deep')

# store S000, SA01, SA02, SA03, SE01, SE02, SE03, SI01, SI02, SI03 columns in a separate dataframe

Origin_Destination_Node_Added_Work = Origin_Destination_Node_Added[['w_geocode', 'h_geocode', 'S000', 'SA01', 'SA02', 'SA03', 'SE01', 'SE02', 'SE03', 'SI01', 'SI02', 'SI03']]

# remove the columns from the original dataframe

# Origin_Destination_Node_Added = Origin_Destination_Node_Added.drop(columns=['S000', 'SA01', 'SA02', 'SA03', 'SE01', 'SE02', 'SE03', 'SI01', 'SI02', 'SI03'])

Origin_Destination_Node_Added

# print memory of Origin_Destination_Node_Added

Origin_Destination_Node_Added.info(memory_usage='deep')

# convert Origin_Destination_Node_Added to array

Origin_Destination_Node_Added_array = Origin_Destination_Node_Added.to_numpy()

Origin_Destination_Node_Added_array

# extract only columns 0 and 1 from Origin_Destination_Node_Added_array

Block_to_Block_Pairs = Origin_Destination_Node_Added_array[:, 0:2]

Block_to_Block_Pairs

# check if each row in Block_to_Block_Pairs array is a unique pair

Block_to_Block_Pairs = Block_to_Block_Pairs.astype(float)  # or int, depending on your data
unique_rows = np.unique(Block_to_Block_Pairs, axis=0)
is_all_unique = len(Block_to_Block_Pairs) == len(unique_rows)

is_all_unique

# extract only columns 2 to 11 from Origin_Destination_Node_Added_array

LODES_info_to_be_adjusted = Origin_Destination_Node_Added_array[:, 2:12]

LODES_info_to_be_adjusted

len(LODES_info_to_be_adjusted)

# Origin_Destination_Node_Added_array[:, 12:14]

block_pairing_count_list = []

for nodes in Origin_Destination_Node_Added_array[:, 12:14]:
    block_pairing_count = len(nodes[0]) * len(nodes[1])
    block_pairing_count_list.append(block_pairing_count)


len(block_pairing_count_list)

for i in range(len(LODES_info_to_be_adjusted)):
    LODES_info_to_be_adjusted[i] = LODES_info_to_be_adjusted[i] / block_pairing_count_list[i]

LODES_adjusted = LODES_info_to_be_adjusted

Node_to_Node_lists = Origin_Destination_Node_Added_array[:, 12:14]

Node_to_Node_lists

Origin_Destination_Node_Added_array[:,:12][0]

from itertools import product

Node_to_Node_pairs = []

for nodes in Node_to_Node_lists:
    Node_to_Node_pairs.extend(product(nodes[0], nodes[1]))

len(Node_to_Node_pairs)

Node_to_Node_pairs

# Export Node_to_Node_pairs to a pickle file

with open('intermediate_files/Node_to_Node_pairs.pickle', 'wb') as handle:
    pickle.dump(Node_to_Node_pairs, handle, protocol=pickle.HIGHEST_PROTOCOL)

    

# Export Origin_Destination_Node_Added_array to a pickle file

with open('intermediate_files/Origin_Destination_Node_Added_array.pickle', 'wb') as handle:
    pickle.dump(Origin_Destination_Node_Added_array, handle, protocol=pickle.HIGHEST_PROTOCOL)

# Export block_pairing_count_list to a pickle file

with open('intermediate_files/block_pairing_count_list.pickle', 'wb') as handle:
    pickle.dump(block_pairing_count_list, handle, protocol=pickle.HIGHEST_PROTOCOL)
