==> 03_igraph_generation_rev.ipynb <==
import pickle
import numpy as np
import pandas as pd
import time
from collections import defaultdict
import random
import igraph as ig

# print("Importing nodes_edges_ucla_access.pickle...")
# with open(r'nodes_edges_ucla_access.pickle', 'rb') as handle:
with open(r'nodes_edges_ucla_big_graph.pickle', 'rb') as handle:
    B_matrix_sliced,B_matrix_str_sliced,nodes_coordinates_array = pickle.load(handle)


B_matrix_sliced

# convert all values in B_matrix_sliced from float to int
B_matrix_sliced = B_matrix_sliced.astype(int)
